requirements.txt
Django==1.11.12
pytz==2018.5
dj-database-url==0.5.0
django-import-export==1.0.1
djangorestframework==3.8.2
gunicorn==19.8.1
Pillow==5.1.0
psycopg2==2.7.4
whitenoise==3.3.1